+++ src/components/HelloWorldWriteUserNameComponent/helloworldwriteusernamecomponent.css
@@ -1,0 +1,9 @@
+.write-user-feature {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    gap: 16px;
+    border: 1px solid black;
+    padding: 0.5rem 15rem;
+}
