+++ src/components/StellarWalletConnection/stellarwalletconnection.css
@@ -4,6 +4,9 @@
   align-items: center;
   justify-content: center;
   gap: 16px;
+  border: 1px solid black;
+  padding: 0.5rem 15rem;
+
 }
 
 .wallet-connected-btn {
