+++ src/index.css
@@ -9,3 +9,9 @@ body {
   font-family: sans-serif;
   min-height: 90vh;
 }
+
+.app {
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+}
